Phase 3: split the DOM glue layer into single-responsibility modules
Pure code movement — no behavior change, proven by the untouched e2e
suite. script.ts (757 lines) becomes:

- dom.ts             el<T>() helper + shared element references
- labels.ts          user-facing texts (single source of truth)
- cloudinary.ts      status image URLs + offline pre-cache
- storage.ts         last-station persistence
- soundEffects.ts    audioInstance (blob preload, warmUp, ensure) + cache
- mediaSession.ts    Now Playing metadata/handlers + iOS re-registration
- serviceWorker.ts   registration + deferred idle-time reload
- theme.ts           prefers-color-scheme theme color
- stationSelector.ts accessible listbox, reports picks via onSelect()
- main.ts            entry point: wiring only, same execution order

Notable seams:
- the core<->mediaSession cycle is resolved by explicit injection
  (initMediaSession before createRadioCore, connectMediaSessionCore
  after) instead of a mutable let;
- mediaSession no longer calls into serviceWorker — main composes
  updateMediaSession + maybeReloadForPendingServiceWorkerUpdate in the
  deps, preserving the original order.

Bundle output unchanged (single js/index.js chunk), so sw.js APP_SHELL
needs no bump. Typecheck clean, 63/63 unit, 37/37 e2e untouched, full
production-preview smoke passing.
