+++ src/index.html
@@ -232,7 +232,7 @@ c1 89 -2 107 -26 157 -33 70 -106 146 -173 178 -56 27 -193 40 -243 22 l-29
     <p class="mt-2"><a href="./downloads/coji-radio-romania.apk" download>Descarcă aplicația Android (0.9)</a></p>
   </footer>
 
-  <script type="module" src="./js/script.ts"></script>
+  <script type="module" src="./js/main.ts"></script>
 </body>
 
 </html>
