+++ src/js/labels.ts
@@ -1,0 +1,6 @@
+/** All user-facing labels (single source of truth). */
+export const LABELS = {
+  appName:  'Coji Radio Player',
+  loading:  'Se încarcă...',
+  error:    'Eroare',
+};
