+++ src/js/theme.ts
@@ -1,0 +1,18 @@
+/** Browser theme color, following the OS light/dark preference. */
+
+function updateThemeColor() {
+  const isDarkMode = window.matchMedia('(prefers-color-scheme: dark)').matches;
+  const themeColor = isDarkMode ? '#434238' : '#fffdef';
+  let metaThemeColor = document.querySelector('meta[name="theme-color"]');
+  if (!metaThemeColor) {
+    metaThemeColor = document.createElement('meta');
+    metaThemeColor.setAttribute('name', 'theme-color');
+    document.head.appendChild(metaThemeColor);
+  }
+  metaThemeColor.setAttribute('content', themeColor);
+}
+
+export function initThemeColor(): void {
+  updateThemeColor();
+  window.matchMedia('(prefers-color-scheme: dark)').addEventListener('change', updateThemeColor);
+}
